requirements.txt
# Core dependencies for Multimodal AI Caption Matcher
# Compatible with Python 3.10, 3.11, 3.12, 3.13

# Deep Learning Framework
torch>=2.0.0,<2.5.0
torchvision>=0.15.0,<0.20.0

# Transformers and CLIP model
transformers>=4.30.0,<5.0.0

# Image processing
Pillow>=10.0.0,<12.0.0

# Scientific computing
numpy>=1.21.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0

# Configuration and utilities
PyYAML>=6.0.0,<7.0.0
tqdm>=4.64.0,<5.0.0

# Optional: For enhanced features
matplotlib>=3.5.0,<4.0.0
pandas>=1.5.0,<3.0.0

# Development dependencies (optional)
jupyter>=1.0.0,<2.0.0
ipykernel>=6.0.0,<7.0.0